MAINT: カスタムフックを使って見やすくした
